Module minicbor::decode

Traits and types for decoding CBOR.

This module defines the trait Decode and the actual Decoder.

Structs

ArrayIter

An iterator over array elements.

BytesIter

An iterator over byte slices.

Decoder

A non-allocating CBOR decoder.

MapIter

An iterator over map entries.

Probe

A decoding probe to to look ahead what comes next.

StrIter

An iterator over string slices.

Enums

Error

Decoding errors.

Traits

Decode

A type that can be decoded from CBOR.