Struct minicbor::decode::Probe

pub struct Probe<'a, 'b> { /* fields omitted */ }

A decoding probe to to look ahead what comes next.

A Probe derefs to Decoder and thus can be used like one without affecting the decoder from which it was created.