Struct minicbor::encode::Encoder

pub struct Encoder<W> { /* fields omitted */ }

A non-allocating CBOR encoder writing encoded bytes to the given Write sink.

Implementations

impl<W: Write> Encoder<W>

pub fn new(writer: W) -> Encoder<W>

Construct an Encoder that writes to the given Write sink.

pub fn into_inner(self) -> W

Get back the Write impl.

pub fn encode<T: Encode>(&mut self, x: T) -> Result<&mut Self, Error<W::Error>>

Encode any type that implements Encode.

pub fn u8(&mut self, x: u8) -> Result<&mut Self, Error<W::Error>>

Encode a u8 value.

pub fn i8(&mut self, x: i8) -> Result<&mut Self, Error<W::Error>>

Encode an i8 value.

pub fn u16(&mut self, x: u16) -> Result<&mut Self, Error<W::Error>>

Encode a u16 value.

pub fn i16(&mut self, x: i16) -> Result<&mut Self, Error<W::Error>>

Encode an i16 value.

pub fn u32(&mut self, x: u32) -> Result<&mut Self, Error<W::Error>>

Encode a u32 value.

pub fn i32(&mut self, x: i32) -> Result<&mut Self, Error<W::Error>>

Encode an i32 value.

pub fn u64(&mut self, x: u64) -> Result<&mut Self, Error<W::Error>>

Encode a u64 value.

pub fn i64(&mut self, x: i64) -> Result<&mut Self, Error<W::Error>>

Encode an i64 value.

pub fn null(&mut self) -> Result<&mut Self, Error<W::Error>>

Encode a CBOR null value.

pub fn undefined(&mut self) -> Result<&mut Self, Error<W::Error>>

Encode a CBOR undefined value.

pub fn simple(&mut self, x: u8) -> Result<&mut Self, Error<W::Error>>

Encode a CBOR simple value.

pub fn f32(&mut self, x: f32) -> Result<&mut Self, Error<W::Error>>

Encode an f32 value.

pub fn f64(&mut self, x: f64) -> Result<&mut Self, Error<W::Error>>

Encode an f64 value.

pub fn bool(&mut self, x: bool) -> Result<&mut Self, Error<W::Error>>

Encode a bool value.

pub fn char(&mut self, x: char) -> Result<&mut Self, Error<W::Error>>

Encode a char value.

pub fn tag(&mut self, x: Tag) -> Result<&mut Self, Error<W::Error>>

Encode a CBOR tag.

pub fn bytes(&mut self, x: &[u8]) -> Result<&mut Self, Error<W::Error>>

Encode a byte slice.

pub fn str(&mut self, x: &str) -> Result<&mut Self, Error<W::Error>>

Encode a string slice.

pub fn array(&mut self, len: u64) -> Result<&mut Self, Error<W::Error>>

Begin encoding an array with len elements.

pub fn map(&mut self, len: u64) -> Result<&mut Self, Error<W::Error>>

Begin encoding a map with len entries.

pub fn begin_array(&mut self) -> Result<&mut Self, Error<W::Error>>

Begin encoding an array of unknown size.

Use Encoder::end to terminate the array.

pub fn begin_bytes(&mut self) -> Result<&mut Self, Error<W::Error>>

Begin encoding an indefinite number of byte slices.

Use Encoder::end to terminate.

pub fn begin_map(&mut self) -> Result<&mut Self, Error<W::Error>>

Begin encoding a map of unknown size.

Use Encoder::end to terminate the map.

pub fn begin_str(&mut self) -> Result<&mut Self, Error<W::Error>>

Begin encoding an indefinite number of string slices.

Use Encoder::end to terminate.

pub fn end(&mut self) -> Result<&mut Self, Error<W::Error>>

Terminate an indefinite collection.

pub fn ok(&mut self) -> Result<(), Error<W::Error>>

Syntactic sugar for Ok(()).