Trait minicbor::encode::Encode

pub trait Encode {
    fn encode<W: Write>(
        &self, 
        e: &mut Encoder<W>
    ) -> Result<(), Error<W::Error>>;
}

A type that can be encoded to CBOR.

If this type’s CBOR encoding is meant to be decoded by Decode impls derived with minicbor_derive it is advisable to only produce a single CBOR data item. Tagging, maps or arrays can and should be used for multiple values.

Required methods

fn encode<W: Write>(&self, e: &mut Encoder<W>) -> Result<(), Error<W::Error>>

Encode a value of this type using the given Encoder.